Make ou and orgUnitPath attrs interchangeable
In some contexts, the connector uses ou and in others orgUnitPath. These can
now be used interchangeably in the opCreate, opModify, and opValidate
operations.

Fixed username so it cannot be all-whitespace. Using a whitespace username
results in a 500 bad request error from google.

Fixed objectClass so it cannot be whitespace.